Split build command host effect tests
