+++ tests/integration/cli_build/build_command_host_effects/env_reads.rs
@@ -1,0 +1,5 @@
+#[path = "env_reads/basic.rs"]
+mod basic;
+
+#[path = "env_reads/ordering.rs"]
+mod ordering;
